TT-678: make stub-rp be runnable on PaaS
One major change is that passport-verify now needs to be located within
the stub-rp directory. This is so that stub-rp is pushed to PaaS the
passport-verify dependency is available. This will be different once we
move past the prototyping phase as the app and library will be located
in seperate git repos