bau: Hide / show SAML form in pure JS
This works around an issue where the user's web server sets a
Content-Security-Policy header which has allows inline JS, but
not inline CSS.

Prior to this commit the form would not be hidden as the JS attempts to
add an inline style, which is disabled by the CSP. By setting the
display property directly in JS we don't need to depend on CSS being
enabled.

I've tested this locally by `yarn link`ing this into
`passport-verify-stub-relying-party` and it still works. Once we've
published a new version to npm we'll need to bump the version in PVSRP.