bau: Allow Level of Assurance 1
Previously passport-verify hardcoded the Level of Assurance to LEVEL_2,
so even though the VSP is capable of supporting both levels services
weren't able to use this for LEVEL_1 without forking passport-verify.

There are two situations where the levelOfAssurance parameter is needed:
* The generateAuthnRequest call (optional - not used by the current
  version of the VSP / Verify)
* The translateResponse call (required - specifies the minimum allowed
  level)

This commit allows the user to specify the level of assurance when they
initialise the strategy and passes the value through to both API calls.

The docs have been updated accordingly.