bau: Pin version of tslint
In node 6 the package.lock file is ignored, so the build is
non-deterministic. We just had an issue where the linter upgraded and
caused the build to break. It doesn't seem like a good idea to have the
linter version changing at random, so let's pin it to 5.10.0.

If we wanted to we could remove this pin once we drop support for node 6.