x1b66RTZ: Renamed the non-matching parts of the code
I think using NonMatching in the class and implementation names
makes it confusing that the thing isn't matching, e.g. NonMatchingAddress.
I went for an "Identity" as a placeholder for non-matching stuff.
It also makes it easier if we ever decide to remove the matching parts of the code
as we'll be only left with "Identity" which is not specific to matching/non-matching.

+ added code comments for the new code
+ updated the README accordingly with some notes on the usage.